// src/types/tesla.ts
interface Timed {
    timestamp: number;
}
export interface Token {
    access_token: string;
    token_type: string;
    expires_in: number;
    refresh_token: string;
    created_at: number;
}

export interface Vehicle {
    id: number;
    vehicle_id: number;
    vin: string;
    display_name: string;
    option_codes: string;
    access_type: string;
    tokens: string[];
    state: 'online' | 'offline' | 'asleep';
    in_service: boolean;
    id_s: string;
    calendar_enabled: boolean;
    api_version: number;
    backseat_token: null;
    backseat_token_updated_at: null;
    vehicle_config: any | null;
}

export interface Climate extends Timed {
    battery_heater: boolean;
    battery_heater_no_power: any;
    climate_keeper_mode: string;
    defrost_mode: number;
    driver_temp_setting: number;
    fan_status: number;
    inside_temp: number;
    is_auto_conditioning_on: boolean;
    is_climate_on: boolean;
    is_front_defroster_on: boolean;
    is_preconditioning: boolean;
    is_rear_defroster_on: boolean;
    left_temp_direction: number;
    max_avail_temp: number;
    min_avail_temp: number;
    outside_temp: number;
    passenger_temp_setting: number;
    remote_heater_control_enabled: boolean;
    right_temp_direction: number;
    seat_heater_left: number;
    seat_heater_right: number;
    side_mirror_heaters: boolean;
    wiper_blade_heater: boolean;
}